App.tsx bifurcated
