+++ src/navigation/AuthNavigator.tsx
@@ -1,0 +1,52 @@
+import React from 'react';
+import { createStackNavigator } from '@react-navigation/stack';
+import { RootStackParamList } from './types';
+import SplashScreen from '../components/auth/splash/SplashScreen';
+import LoginScreen from '../components/auth/Login/LoginScreen';
+import OnboardingScreen from '../components/auth/onboarding/OnboardingScreen';
+import Logo from '../assets/images/AnantaLogo.svg';
+
+import OtpScreen from '../components/auth/otp/OtpScreen';
+const Stack = createStackNavigator<RootStackParamList>();
+
+export const AuthNavigator = () => {
+  return (
+    <Stack.Navigator
+      initialRouteName="Splash"
+      screenOptions={{ headerShown: false }}
+    >
+      <Stack.Screen name="Splash" component={SplashScreen} />
+      <Stack.Screen
+        name="Onboarding"
+        component={OnboardingScreen}
+        options={{ gestureEnabled: false }}
+      />
+      <Stack.Screen
+        name="Login"
+        component={LoginScreen}
+        options={{
+          gestureEnabled: false,
+          headerShown: true,
+          headerTitle: () => <Logo width={100} height={100} />,
+          headerTitleAlign: 'center',
+          headerStyle: { backgroundColor: 'black' },
+          headerTintColor: '#D4AF37',
+          headerLeft: () => null,
+        }}
+      />
+      <Stack.Screen
+        name="Otp"
+        component={OtpScreen}
+        options={{
+          gestureEnabled: true,
+          headerShown: true,
+          headerTitle: () => <Logo width={100} height={100} />,
+          headerTitleAlign: 'center',
+          headerStyle: { backgroundColor: 'black' },
+          headerTintColor: '#D4AF37',
+          headerBackButtonDisplayMode: 'minimal',
+        }}
+      />
+    </Stack.Navigator>
+  );
+};
