+++ src/navigation/TabNavigator.tsx
@@ -1,0 +1,121 @@
+import React from 'react';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { TabParamList } from './types';
+import HomeScreen from '../screens/HomeScreen';
+import BookingScreen from '../screens/BookingScreen';
+import ServicesScreen from '../screens/ServicesScreen';
+import ProfileScreen from '../screens/ProfileScreen';
+import { useNavigation } from '@react-navigation/native';
+import { DrawerActions } from '@react-navigation/native';
+import { TouchableOpacity } from 'react-native';
+import Logo from '../assets/images/AnantaLogo.svg';
+import MenuIcon from '../assets/images/MenuIcon.svg';
+import Icon from 'react-native-vector-icons/Ionicons';
+
+// Extend the TabParamList if needed
+declare global {
+  namespace ReactNavigation {
+    interface RootParamList extends TabParamList {}
+  }
+}
+
+const Tab = createBottomTabNavigator<TabParamList>();
+
+export const TabNavigator = () => {
+  const navigation = useNavigation();
+
+  return (
+    <Tab.Navigator
+      screenOptions={({ route }) => ({
+        tabBarIcon: ({ focused, color, size }) => {
+          let iconName: string;
+
+          switch (route.name) {
+            case 'Home':
+              iconName = focused ? 'home' : 'home-outline';
+              break;
+            case 'Booking':
+              iconName = focused ? 'calendar' : 'calendar-outline';
+              break;
+            case 'Services':
+              iconName = focused ? 'list' : 'list-outline';
+              break;
+            case 'Profile':
+              iconName = focused ? 'person' : 'person-outline';
+              break;
+            default:
+              iconName = 'home-outline';
+          }
+
+          return <Icon name={iconName} size={size} color={color} />;
+        },
+        tabBarActiveTintColor: '#D4AF37',
+        tabBarInactiveTintColor: '#666666',
+        tabBarStyle: {
+          backgroundColor: '#000000',
+          borderTopWidth: 1,
+          borderTopColor: '#333333',
+          height: 60,
+          paddingBottom: 8,
+          paddingTop: 8,
+        },
+        tabBarLabelStyle: {
+          fontSize: 12,
+          fontWeight: '500',
+        },
+        headerStyle: {
+          backgroundColor: '#000000',
+        },
+        headerTintColor: '#D4AF37',
+        headerTitleStyle: {
+          fontWeight: 'bold',
+        },
+      })}
+    >
+      <Tab.Screen
+        name="Home"
+        component={HomeScreen}
+        options={{
+          title: 'Ananta Resorts',
+          headerShown: true,
+          headerTitle: () => <Logo width={100} height={100} />,
+          headerTitleAlign: 'center',
+          headerStyle: { backgroundColor: 'black' },
+          headerTintColor: '#D4AF37',
+          headerLeft: () => (
+            <TouchableOpacity
+              onPress={() => navigation.dispatch(DrawerActions.toggleDrawer())}
+              style={{ marginLeft: 15 }}
+            >
+              <MenuIcon width={30} height={30} />
+            </TouchableOpacity>
+          ),
+        }}
+      />
+      <Tab.Screen
+        name="Booking"
+        component={BookingScreen}
+        options={{
+          title: 'Book Stay',
+          headerShown: true,
+        }}
+      />
+      <Tab.Screen
+        name="Services"
+        component={ServicesScreen}
+        options={{
+          title: 'Services',
+          headerShown: true,
+        }}
+      />
+      <Tab.Screen
+        name="Profile"
+        component={ProfileScreen}
+        options={{
+          title: 'Profile',
+          headerShown: true,
+        }}
+      />
+    </Tab.Navigator>
+  );
+};
